src/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
base58==2.1.1
cffi==1.17.1
click==8.1.8
colorama==0.4.6
confluent-kafka==2.8.2
cryptography==44.0.2
dotenv==0.9.9
fastapi==0.115.11
h11==0.14.0
idna==3.10
keyboard==0.13.5
mysql-connector-python==9.2.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
python-multipart==0.0.20
setuptools==76.1.0
sniffio==1.3.1
starlette==0.46.1
typing_extensions==4.12.2
uvicorn==0.34.0
wheel==0.45.1
